Extract reusable CoreData helpers into a separate file
